+++ test/fixtures/parse/require/User.js
@@ -1,0 +1,6 @@
+import { t } from '../../../../.';
+
+export default t.struct({
+  name: t.String,
+  surname: t.String
+}, 'User');